+++ src/index.ts
@@ -1,1 +1,3 @@
-export { default as useDropdown } from './hooks/useDropdown'
+export { default as useDropdown } from './hooks/useDropdown';
+
+export { default as SelectDropdown } from './components/SelectDropdown';
